refactor/809 - Consolidate constants to routeConstants and waypointConstants
